LayoutTests/media/media-rvfc-paused-offscreen-mp4-expected.txt

EVENT(loadedmetadata)
EXPECTED (0 == '0') OK
EXPECTED (0 == '0') OK
EXPECTED (320 == '320') OK
EXPECTED (240 == '240') OK
EXPECTED (6.009 == '6.009') OK
EXPECTED (6.009 == '6.009') OK
EXPECTED (320 == '320') OK
EXPECTED (240 == '240') OK
